Tugas Pertemuan 12

+++ lib/model/cuaca_model.dart
@@ -11,7 +11,7 @@ class CuacaModel {
   int? timezone;
   int? id;
   String? name;
-  double? cod;
+  int? cod;
 
   CuacaModel(
       {this.coord,
